requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
appnope==0.1.3
    # via ipython
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via pytest
backcall==0.2.0
    # via ipython
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
exceptiongroup==1.1.0
    # via pytest
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via
    #   -r requirements.in
    #   pytest-factoryboy
faker==17.0.0
    # via factory-boy
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.0.0
    # via pytest
ipdb==0.13.11
    # via -r requirements.in
ipython==8.10.0
    # via
    #   -r requirements.in
    #   ipdb
jedi==0.18.2
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
packaging==23.0
    # via pytest
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.36
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.14.0
    # via ipython
pytest==7.2.1
    # via pytest-factoryboy
pytest-factoryboy==2.5.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
stack-data==0.6.2
    # via ipython
tomli==2.0.1
    # via
    #   ipdb
    #   pytest
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.5.0
    # via pytest-factoryboy
wcwidth==0.2.6
    # via prompt-toolkit
